Added a 'mk'-style interface to rotation queries
This allows them to be somewhat polymorphic with all the other query functions which have a 'mk'/'exec' pair where the 'mk' creates an Allocator and the 'exec' executes the query. However, the rotation query 'mk' function creates an (Allocator, Transformer, Complete) tuple, suitable for use with the generic 'query' function.